requirements.txt
httpx
python-dotenv
torch
transformers
scikit-learn
datasets
streamlit
colorama
